Updated tunings and parameters
Better reflect Husky shape.

Reduced size of track and search space so it better represents MUR race car ratio (car width of 1.4m to min track width of 3.5m). Probably need some calc to verify it this ratio, but nonetheless, thinner tracks.

It's also a bit more of a challenge.